pyproject.toml
[tool.poetry]
name = "sift"
version = "0.1.0"
description = "An experimental tool to build a knowledge graph from a code repository"
authors = ["Frank Goortani"]
readme = "README.md"
packages = [
    { include = "src", from = "." }
]

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.41.0"
gitpython = "^3.1.43"
requests = "^2.32.3"
rich = "^13.7.1"
python-dotenv = "^1.0.1"
asyncio = "^3.4.3"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.1"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
mypy = "^1.11.1"
black = "^24.8.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
select = ["E", "W", "F"]
ignore = ["E203", "E266", "E501"]

[tool.poetry.scripts]
sift = "src.cli:main"
